backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
sqlalchemy>=2.0.25
pydantic>=2.6.0
python-dotenv>=1.0.0
email-validator>=2.1.0